requirements.txt
asgiref==3.2.3
astroid==2.3.3
boto3==1.12.46
botocore==1.15.46
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
confusable-homoglyphs==3.2.0
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-forms-bootstrap==3.1.0
django-html5==1.0.0
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-registration==3.1
django-storages==1.9.1
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-S3==0.3.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
pymongo==3.10.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.2
six==1.13.0
sqlparse==0.3.0
stripe==2.43.0
urllib3==1.25.8
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2
